YamahaMX/Reaper/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Work\MusicConfigAndScriptRepository\YamahaMX\Reaper\YamahaMXMidiUtils.py: 2
mido == 1.2.9

# D:\Work\MusicConfigAndScriptRepository\YamahaMX\Reaper\YamahaMXMidiUtils.py: 1
# D:\Work\MusicConfigAndScriptRepository\YamahaMX\Reaper\YamahaMXReaperLoadFirstVoiceOnTrack.py: 1
# D:\Work\MusicConfigAndScriptRepository\YamahaMX\Reaper\YamahaMXReaperUtils.py: 1
python_reapy == 0.7.2
